Move hasWebDav tests to repository
